OK, make opentofu work with external helper.
This was more complicated than I hoped for.
We could have done some HCL magic IF the openstack provider would have
allowed to query image properties by IDs which it does not seem to.
So the external python3 program is easiest and we reuse what we already
have.

Also comment on heat.

Signed-off-by: Kurt Garloff <kurt@garloff.de>